+++ Changelog.md
@@ -14,6 +14,7 @@
 - Fix: Reinforcing a placed end rod will briefly destroy the end rod before it is replaced with a reinforced end rod
 - Fix: Crash when trying to toggle the redstone state of a camera immediately after mounting it
 - Fix: Crash when trying to remove the passcode of a Briefcase using a Universal Key Changer
+- Fix: The Display Case doesn't drop inserted modules when the block the display case is placed on is removed
 
 --------------------------Changelog for v1.9.9 of SecurityCraft--------------------------